Changed player turn management
