--- Assets/Scenes/New Grid Test Scene_Profiles.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f658406838735b04a9f793913f9f891b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
