+++ Assets/Scripts/Battlefield/AbstractTurnBrain.cs
@@ -4,13 +4,17 @@ using UnityEngine;
 using SwordAndBored.Battlefield.CreaturScripts;
 using SwordAndBored.Battlefield.TurnMechanism;
 using SwordAndBored.Battlefield.CameraUtilities;
+using Cinemachine;
 
 
 namespace SwordAndBored.Battlefield
 {
     public abstract class AbstractTurnBrain : MonoBehaviour
     {
-        public TurnManager turnManager;
-        abstract public void UpdateBrainHandler();
+        public abstract void DoTurn();
+        public abstract string GetName();
+        public abstract bool GetTurnEnd();
+        public abstract CinemachineVirtualCamera GetCam();
+        public abstract bool HasActionLeft();
     }
 }
