--- Assets/Scripts/Battlefield/PlayerInputHandler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0cfbe605db6a7c74a82436b6724fa5fa
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
