+++ Assets/Scripts/Battlefield/TurnMechanism/TurnManager.cs
@@ -12,7 +12,7 @@ namespace SwordAndBored.Battlefield.TurnMechanism
     {
         [Header("Units")]
         public GameObject[] units;
-        public UniqueCreature activePlayer;
+        public AbstractTurnBrain activePlayer;
         [Header("UI")]
         TurnOrderController manager;
         public TextMeshProUGUI text;
@@ -22,20 +22,26 @@ namespace SwordAndBored.Battlefield.TurnMechanism
         void Start()
         {
             manager = new TurnOrderController(units, new RandomShuffler<GameObject>());
-            activePlayer = manager.NextEntity().GetComponent<UniqueCreature>();
-            text.text = "Current Player: " + activePlayer.creatureName;
+            activePlayer = manager.NextEntity().GetComponent<AbstractTurnBrain>();
+            text.text = "Current Player: " + activePlayer.GetName();
+            activePlayer.DoTurn();
         }
 
         public void nextTurn()
         {
-            activePlayer = manager.NextEntity().GetComponent<UniqueCreature>();
-            text.text = "Current Player: " + activePlayer.creatureName;
-            activePlayer.StartTurn();
+            activePlayer = manager.NextEntity().GetComponent<AbstractTurnBrain>();
+            text.text = "Current Player: " + activePlayer.GetName();
+            activePlayer.DoTurn();
         }
 
         void Update()
         {
-            if (activePlayer.action)
+            if (activePlayer && !activePlayer.GetTurnEnd())
+            {
+                nextTurn();
+            }
+
+            if (activePlayer.HasActionLeft())
             {
                 actionsLeft[0].color = new Color(0, 0, 1, .2f);
             }
