--- Assets/TestConn.db.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 49b161a41b9e425429dceff03d0e2d3a
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
